Cache some blueprint fields as well
